% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\docType{data}
\name{epa}
\alias{epa}
\title{EPA}
\format{
A data frame with 100 rows and 1 column:
\describe{
  \item{MPG}{miles per gallon}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
epa
}
\description{
A data set containing the MPG of 100 vehicles
}
\keyword{datasets}
