% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybiplot.R
\name{mybiplot}
\alias{mybiplot}
\title{mybiplot}
\usage{
mybiplot(x = double(), k = double())
}
\arguments{
\item{x}{a numeric vector}

\item{k}{the number of standard deviations to compare with}
}
\value{
a list
}
\description{
a function which assigns each value in a vector as "in", "outu", or "outl" based on thier distance from the mean
}
\examples{
\dontrun{mybiplot(x = mpg, k=1)}
}
