% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.math4753biplot.R
\name{print.math4753biplot}
\alias{print.math4753biplot}
\title{print.math4753biplot}
\usage{
\method{print}{math4753biplot}(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{passes extra options into the function}
}
\value{
a histogram
}
\description{
takes a numeric vector and plots a histogram using ggploit2
}
\examples{
\dontrun{print.math4753biplot(x = mpg)}
}
