% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.math4753biplot.R
\name{summary.math4753biplot}
\alias{summary.math4753biplot}
\title{summary.math4753biplot}
\usage{
\method{summary}{math4753biplot}(object, ...)
}
\arguments{
\item{object}{a numeric vector}

\item{...}{additional options}
}
\value{
a summary of mean, median, IQR, n grouped by their .data$inout status
}
\description{
produces a summary which includes mean, median, and IQR
}
\examples{
\dontrun{summarymath4753biplot(x = mpg)}
}
